Week_09-10_Projects_and_Advanced_Topics/Projects/ToDo_App/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.3
tenacity==8.2.3